Create Sum of Series.cpp
m is assigned the value of n. This is done to make the code more readable and to avoid multiple occurrences of n in the formula.
s is then calculated using the formula for the sum of an arithmetic series: 

Finally, s is returned, which holds the sum of the first n natural numbers.